Moved notification to different page
