+++ auth-server/src/resolvers/mutation/changeNotificationPreference.ts
@@ -1,18 +1,18 @@
 import getTokenFromReq from '../../utils/getTokenFromReq';
 import AuthService from '../../services/auth';
-import { Context, MessageType, NotificationPreferencesType } from '../../types';
+import { Context, ChangeResponseType, NotificationPreferencesType } from '../../types';
 import messages from '../../utils/messages';
 
 interface argsType {
 	notificationPreferences: NotificationPreferencesType
 }
 
-export default async (parent, { notificationPreferences }: argsType, ctx: Context): Promise<MessageType> => {
+export default async (parent, { notificationPreferences }: argsType, ctx: Context): Promise<ChangeResponseType> => {
 	const token = getTokenFromReq(ctx.req);
 	const { postParticipated, postCreated, newProposal, ownProposal } = notificationPreferences;
 
 	const authServiceInstance = new AuthService();
-	await authServiceInstance.ChangeNotificationPreference(
+	const updatedJWT = await authServiceInstance.ChangeNotificationPreference(
 		token,
 		{
 			postParticipated,
@@ -22,5 +22,5 @@ export default async (parent, { notificationPreferences }: argsType, ctx: Contex
 		}
 	);
 
-	return { message: messages.NOTIFICATION_PREFERENCE_CHANGE_SUCCESSFUL };
+	return { message: messages.NOTIFICATION_PREFERENCE_CHANGE_SUCCESSFUL, token: updatedJWT };
 };
